requirements.txt
Pillow
python-dotenv
img2pdf
tqdm
